requirements.txt
pandas
numpy
scikit-learn
tensorflow
pyyaml
joblib
flask
streamlit
kagglehub